get nearby zip codes
